move files into grid
