+++ src/features/references/editor/ReferencesTableView.tsx
@@ -15,8 +15,8 @@ import { ReferenceItem, ReferenceItemStatus } from '../../../types/ReferenceItem
 import { ReferencesItemStatusLabel } from '../components/ReferencesItemStatusLabel';
 import { TopActionIcon } from '../components/TopActionIcon';
 import { UploadTipInstructions } from '../components/UploadTipInstructions';
-import { loadColumnsState, resetColumnsState, saveColumnsState } from './columnsState';
 import { authorsFormatter, firstAuthorFormatter } from './formatters';
+import { loadColumnsState, resetColumnsState, saveColumnsState } from './grid/columnsState';
 
 export function ReferencesTableView({ defaultFilter = '' }: { defaultFilter?: string }) {
   const references = useAtomValue(getReferencesAtom);
